Preparation for VES Avatar and VES Mask (scenes, prefabs, scripts)

+++ VR/Assets/XROSUI/Scripts/UI/SystemMenu/Slider_Audio.cs
@@ -9,8 +9,8 @@ public class Slider_Audio : MonoBehaviour
     public AudioMixer mixer;
     public ENUM_Audio_Type type;
 
-    public void SetVolume(int level)
+    public void SetVolume(float level)
     {
-        Core.Ins.AudioManager.SetVolume(level, type);
+        Core.Ins.AudioManager.SetVolume((int)level, type);
     }
 }
